+++ test/xLtest.c
@@ -81,7 +81,7 @@ main(Void)
 {
 	char *s, *s1, *se, *se1;
 	int dItry, i, ndig = 0, r = 1;
-	union { long double d; ULong bits[3]; } u, v[2];
+	union { long double d; uint32_t bits[3]; } u, v[2];
 
 	while(s = fgets(ibuf, sizeof(ibuf), stdin)) {
 		while(*s <= ' ')
@@ -102,11 +102,11 @@ main(Void)
 		  case '#':
 			/* sscanf(s+1, "%lx %lx %lx", &u.bits[_0],	*/
 			/* 	&u.bits[_1], &u.bits[_2]);		*/
-			u.bits[_0] = (ULong)strtoul(s1 = s+1, &se, 16);
+			u.bits[_0] = (uint32_t)strtoul(s1 = s+1, &se, 16);
 			if (se > s1) {
-			    u.bits[_1] = (ULong)strtoul(s1=se, &se, 16);
+			    u.bits[_1] = (uint32_t)strtoul(s1=se, &se, 16);
 			    if (se > s1)
-				u.bits[_2] = (ULong)strtoul(s1=se, &se, 16);
+				u.bits[_2] = (uint32_t)strtoul(s1=se, &se, 16);
 			    }
 			printf("\nInput: %s", ibuf);
 			printf(" --> f = #%lx %lx %lx\n", u.bits[_0],
